further improvements

+++ src/org/igsq/igsqbot/handlers/CommandHandler.java
@@ -3,7 +3,9 @@ package org.igsq.igsqbot.handlers;
 import io.github.classgraph.ClassGraph;
 import io.github.classgraph.ClassInfo;
 import io.github.classgraph.ScanResult;
+import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.entities.ChannelType;
+import net.dv8tion.jda.api.entities.GuildChannel;
 import net.dv8tion.jda.api.entities.Message;
 import net.dv8tion.jda.api.entities.MessageChannel;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
@@ -11,6 +13,7 @@ import org.igsq.igsqbot.Common;
 import org.igsq.igsqbot.objects.EmbedGenerator;
 import org.igsq.igsqbot.objects.Command;
 import org.igsq.igsqbot.objects.Context;
+import org.igsq.igsqbot.util.Array_Utils;
 
 import java.awt.*;
 import java.util.HashMap;
@@ -23,7 +26,7 @@ public abstract class CommandHandler
 		//Overrides the default, public, constructor
 	}
 
-	private static final ClassGraph CLASS_GRAPH = new ClassGraph().acceptPackages("org.igsq.igsqbot.improvedcommands");
+	private static final ClassGraph CLASS_GRAPH = new ClassGraph().acceptPackages(Common.COMMAND_PACKAGE);
 	private static final Map<String, Command> COMMANDS = new HashMap<>();
 
 	static
@@ -50,7 +53,7 @@ public abstract class CommandHandler
 		final String issuedCommand = (content.contains(" ") ? content.substring(0, content.indexOf(' ')) : content).toLowerCase();
 		final Command cmd = COMMANDS.get(issuedCommand);
 		final MessageChannel channel = event.getChannel();
-		final String[] args = Common.depend(event.getMessage().getContentRaw().split(" "), 0);
+		final String[] args = Array_Utils.depend(event.getMessage().getContentRaw().split(" "), 0);
 
 		if(!message.getContentRaw().startsWith(Common.BOT_PREFIX) || event.getAuthor().isBot())
 		{
@@ -59,7 +62,7 @@ public abstract class CommandHandler
 
 		if(cmd == null)
 		{
-			new EmbedGenerator(channel).text("Command " + issuedCommand + " was not found")
+			new EmbedGenerator(channel).text("Command `" + issuedCommand + "` was not found")
 			.color(Color.RED)
 			.sendTemporary();
 			return;
@@ -69,7 +72,11 @@ public abstract class CommandHandler
 		{
 			return;
 		}
-		event.getMessage().delete().queue();
+
+		if(event.getChannelType().equals(ChannelType.TEXT) && event.getGuild().getSelfMember().hasPermission((GuildChannel) channel, Permission.MESSAGE_MANAGE))
+		{
+			event.getMessage().delete().queue();
+		}
 		Common.commandExecutor.submit(() -> cmd.execute(args, new Context(event)));
 	}
 }
